requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pikepdf==9.5.1
pillow==11.1.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
python-dotenv==1.0.1